Move board tap request model into Models
